Tests: move Mock classes to Fixtures directory
... so all fixtures are grouped together in one directory.
